UI added
Added some basic UI to help play testers.
Fixed some issues in the invoke freeze as well as switched as made all the coroutines use timers so that UI timer elements could be added..
Same changes where made to the revers and slow scripts.
Created a basic script for an object with a rigidbody to be frozen.